Server/requirements.txt
alembic==1.13.2
apturl==0.5.2
astroid==2.9.3
asttokens==2.4.1
async-timeout==5.0.1
Babel==2.8.0
bcrypt==3.2.0
blinker==1.8.2
Brlapi==0.8.3
cachelib==0.13.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.2.2
executing==2.1.0
fasteners==0.14.1
filelock==3.16.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
future==0.18.2
greenlet==3.1.0
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.27.0
isort==5.6.4
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==0.0.0
lockfile==0.12.2
logilab-common==1.8.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.6.1
monotonic==1.6
more-itertools==8.10.0
msgspec==0.18.6
mypy-extensions==0.4.3
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==24.1
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
pipenv==2024.0.2
platformdirs==4.3.3
pluggy==1.5.0
prompt_toolkit==3.0.47
protobuf==3.12.4
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycups==2.0.1
Pygments==2.18.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pylint==2.12.2
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==8.3.2
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==5.2.0
reportlab==3.6.8
requests==2.32.3
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
SQLAlchemy==2.0.34
stack-data==0.6.3
systemd-python==234
toml==0.10.2
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.26.4
virtualenv-clone==0.3.0
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.13.3
xdg==5
xkit==0.0.0
zipp==1.0.0
